requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-frontmatter>=1.0.0
markdown>=3.5.0
watchdog>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
